FOV calculates correctly now
FOV adjustments, near and far plane, and the camera preview only shows if you have the camera selected

+++ Engine/Source/LoadModel.cpp
@@ -116,7 +116,7 @@ void LoadModel::CreatingModel(JsonParsing& json, JSON_Array* array, GameObject* 
 		for (int j = 0; j < s; ++j)
 		{
 			JsonParsing component = parsing.GetJsonArrayValue(arr, j);
-			switch ((ComponentType)component.GetJsonNumber("Type"))
+			switch ((ComponentType)(int)component.GetJsonNumber("Type"))
 			{
 			case ComponentType::MESH_RENDERER:
 			{
